works

+++ db/migrate/20161114071027_create_keys.rb
@@ -1,7 +1,7 @@
 class CreateKeys < ActiveRecord::Migration[5.0]
   def change
     create_table :keys do |t|
-      t.integer :offset
+      t.float :offset
 
       t.timestamps
     end
